requirements.txt
Django==4.2.2
python-dotenv~=0.21.0
gunicorn==20.1.0
psycopg2-binary==2.9.6
setuptools==67.8.0
django-bootstrap4==23.1
google-api-python-client==2.88.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
python-telegram-bot==20.3
celery==5.3.1
redis==4.5.5
django-celery-results==2.5.1
django-celery-beat==2.5.0

requests==2.28.2